Extract usePaginatedQuery — collapse 3 near-identical pagination hooks
useProcessInstancesPage, useAgentRunsPage, and useNamespaceAuditEventsPage
all wrapped useInfiniteQuery with the same shape: same retry/refetchInterval
config, same flatMap-over-pages item extraction, same
{data, loading, error, hasMore, loadingMore, loadMore} return. Extract
that into a single usePaginatedQuery(queryKey, queryFn, selectItems,
enabled) hook; each call site now only supplies its own query key, page
fetcher, and how to pull the item array out of its page shape (`runs`
vs `events` field names differ per endpoint).

Flagged during #1126's self-review (Standards + Big-Picture axes —
meets the review checklist's own "same shape repeated >=3x -> extract"
bar). Tracked as issue #1128 before this commit implements it.
