+++ packages/platform-ui/src/hooks/use-process-instances-page.ts
@@ -1,12 +1,9 @@
 'use client';
 
-import { useMemo } from 'react';
-import { useInfiniteQuery } from '@tanstack/react-query';
-import type { ProcessInstance, WorkflowDisplayStatus } from '@mediforce/platform-core';
+import type { WorkflowDisplayStatus } from '@mediforce/platform-core';
 import { mediforce } from '@/lib/mediforce';
 import { queryKeys } from '@/lib/query-keys';
-import { stopRetryOn4xx } from '@/lib/retry';
-import { STANDARD_LIVE_INTERVAL_MS } from '@/lib/polling-cadence';
+import { usePaginatedQuery } from './use-paginated-query';
 
 const PAGE_SIZE = 20;
 
@@ -34,43 +31,24 @@ export function useProcessInstancesPage(params: {
   const { namespace, workflowFilter, dryRun, archived, displayStatus } = params;
   const enabled = namespace.length > 0;
 
-  const query = useInfiniteQuery({
+  return usePaginatedQuery({
     queryKey: queryKeys.runs.page(namespace, {
       workflow: workflowFilter,
       dryRun,
       archived,
       displayStatus: displayStatus ?? undefined,
     }),
-    queryFn: async ({ pageParam }) =>
+    queryFn: (cursor) =>
       mediforce.runs.listPage({
         namespace,
         workflow: workflowFilter,
         dryRun,
         archived,
         displayStatus: displayStatus ?? undefined,
-        cursor: pageParam,
+        cursor,
         limit: PAGE_SIZE,
       }),
-    initialPageParam: undefined as string | undefined,
-    getNextPageParam: (lastPage) => lastPage.nextCursor,
+    selectItems: (page) => page.runs,
     enabled,
-    retry: stopRetryOn4xx,
-    refetchInterval: (q) => (q.state.error !== null ? false : STANDARD_LIVE_INTERVAL_MS),
   });
-
-  const items = useMemo<ProcessInstance[]>(
-    () => query.data?.pages.flatMap((p) => p.runs) ?? [],
-    [query.data],
-  );
-
-  return {
-    data: items,
-    loading: enabled && query.isPending,
-    error: (query.error as Error | null) ?? null,
-    hasMore: query.hasNextPage,
-    loadingMore: query.isFetchingNextPage,
-    loadMore: () => {
-      void query.fetchNextPage();
-    },
-  };
 }
